Php4 compatibility + some other littele features

+++ tasks/files.php
@@ -12,7 +12,7 @@ switch($params[0])
         $filesLocations = implode(" ",$config["prod_server"]["files"]);
         $tgzFilePath = $tmpDirectory."/files.tgz";
         $cmd[] = "cd ".$projectRootDir;
-        $cmd[] = sprintf('tar -zcf %s %s', $tgzFilePath, $filesLocations);
+        $cmd[] = sprintf('tar -zchf %s %s', $tgzFilePath, $filesLocations);
         $status = shell_exec(implode(";", $cmd));
         // on ne passe pas par un readfile pour retourner les data à cause des limitations de mémoire de php sous apache
         header("Location: http://"._CONST_PROD_DOMAIN.$config["prod_server"]["http_path"]."/".$rootName."/".$tmpDirectoryName."/files.tgz");
@@ -21,9 +21,17 @@ switch($params[0])
         $tgzFilePath = $tmpDirectory."/files.tgz";
         $filesLocations = implode(" ",$config["prod_server"]["files"]);
 
+        $confirm = ask(sprintf("Production files will be downloaded and expanded here %s, continue ? [y]: ", $filesLocations), true);
+        if($confirm != "y") {
+            echo "exiting...";
+            exit;
+        }
+
         $cmd = sprintf('wget "http://'._CONST_PROD_DOMAIN.$config["prod_server"]["http_path"]."/".$rootName.'/task.php?run=files get_prod_export" -O %s',$tgzFilePath);
         $status = taskExecute($cmd,"Downloading from production tgz archive of files ".$filesLocations);
         echo "\nExpanding archive...";
+
+        $cmd = array();
         $cmd[] = "cd ".$projectRootDir;
         $cmd[] = sprintf('tar -zxf %s',$tgzFilePath);
         $status = shell_exec(implode(";", $cmd));
@@ -42,6 +50,6 @@ function help()
   global $rootName;
   echo "Use 'get_prod_export' or 'get_import' argument. \n";
   echo "'get_prod_export' give you a tarball shared files and must be called via http, not in console mode. To call it : http://"._CONST_PROD_DOMAIN.$config["prod_server"]["http_path"]."/".$rootName."/task.php?run=files get_prod_export. \n";
-  echo "'get_import' automatically get the tarball of shared files and expand it into your local project. \n";
+  echo "'prod_import' automatically get the tarball of shared files and expand it into your local project. \n";
 
 }
